+++ CHANGELOG.md
@@ -11,10 +11,13 @@ and the release workflow reads it to set github's release notes.
 
 ## Unreleased
 
+
 ### Added 
 
-- The `/approve` endpoint POST there to approve a peer
-- The `/register` endpoint, used to register a user
+- A WebRTC endpoint, `/we`, for admin commands:
+ - register <email> <peer_name>
+ - verify <target_cingerprint> <otp>
+ - ping [otp]
 
 ## [1.2]